Create Project 1: Orientation Quantification
Project #1: Determining the orientation of the various nuclei on the DAPI channels of the slides with respect to the direction of variable laminar fluid flow (the horizontal axis) with a simple least squares approach. 

First, I preprocessed the relevant slides to create an unambiguous base layer of the shapes of the nuclei to run the preprocessing. As the DAPI slides were grayscale tiffs (Extracted from ZEN format), I converted the images to bitmaps such that all pixels above a variable threshold (32 out of 255 in this case) would be mapped to an “on value” (Ex: 255) while all other values would be converted to an “off value” (Ex: 0). This removes the slight blur that exists at the periphery of the DAPI stain to remove ambiguity about the nuclei borders. 
My approach on this project was to first devise a way to count the number of nuclei that were on a particular slide, in a way such that one nucleus would be counted at a time.
My solution is reminiscent of Monte Carlo methods, as it takes a random point on the base nuclei layer as an argument and looks at its binary value. If the value is an “off value”, then the pixel selected is not part of a nucleus, and is thus ignored. If the value is an “on value”, then the pixel selected is part of a nucleus, and thus undergoes a flood fill over all of the “on values” from that point, which fills the area of the nucleus of the selected point. The nucleus counter ticks up by one, and all the values that have been covered with the flood fill have been tagged such that future calls over the tagged pixels no longer initiate flood fills or counts.This subroutine is looped for a number of iterations that corresponds to the area of the image, such that a single nucleus is categorized at a time.
The flood fill used to fully categorize the nucleus is a variant of the vertical span fill, such that it returns the pixel coordinates of the edges of the structure. The modified floodfill cannot track edges that form convex angles that have a tangent that is greater than 45 degrees from the horizontal axis of the image. However, given the nature of nuclei and its alignment to flow, there should be no such convex angles that exist on the horizontal on the nucleus [1]. This 1D fill is contained in the Vertical function and coordinate gathering is contained in the Vertcaps function. The Horizontal function calls the Vertical and Vertcaps functions along the horizontal component of the nucleus. These functions tag the selected nuclear structure and cover all of the “on” values with a recognizable tag ((255,0,0) on a color .tiff layer generated from the initial structures.
The center of the nucleus (mean of x and y coordinates) are calculated and used to find the slope and constant parameters of a simple linear regression. The slope is calculated as the x values minus the x mean, multiplied by the y values minus the y mean all summed together, which is all divided by the sum of all of the x values minus the x mean. This parameter, represented by the variable b, can also be represented as the sample standard deviation of variables x and y divided by the sample standard deviation of x squared. The constant parameter is thus calculated by the slope multiplied by the x mean, all subtracted from the y mean. After calculated, a line constructed from the slope and constant would be overlaid over the center of the nucleus, and the length of the line segment of the line that lies on the nucleus would be calculated. In addition, another line is constructed from the same constant, but finding the negative inverse of the slope, which is also used to derive another line segment. The numerical lengths of each of these line segments are calculated, with the larger segment being designated as the major axis, while the smaller segment is denoted as the minor axis. These two perpendicular axes best represent the pseudo-ellipsoid shape of the nucleus, and are the numerical values that are used to determine the orientation of the nuclei on the slide. The longer of these two axes is laid over the corresponding nucleus in figure 1.3, where purple axes represent horizontal alignments and green axes represent vertical alignments. This processing is all encompassed in the lines function. It is worth noting that the lines function works remarkably well for the given nuclear shape, as it accommodates ellipsoids that can deviate from strict ellipses to a significant degree.