Project #2: Segmenting cells using Calcium/Calvin channels.
It should be noted that the Cavin channels isolate and show caveolae specific proteins on the surfaces of cells. Thus, areas highlighted by the cavin channels tend to exclude areas that cell membranes run through on the full microscope image containing all of the constituent channels. In addition, the cell membranes of the endothelial cells were identified as various thin lines in the calcium dye channel, as opposed to the BAPTA-1-AM ester gradient that is present in the center of the cells in the dye channel. The approach I took was to isolate the cell membrane “signal” in the endothelial cells in the dye channel by removing “noise” such as the BAPTA-1-AM ester gradient. To do so, I used spatial distribution information that was present in the CAVIN channel to isolate the cell membrane.
First, I created two images based on the CAVIN and dye channels, where the CAVIN channel was represented as the gaussian blur of the tiff of the CAVIN channel with “on” and “off” values as described above (variable threshold was ~⅓ of pixel of maximum value). The dye image was a grayscale tiff of the dye channel for that particular image. The images were converted to matrices, and numpy commands were used to subtract the CAVIN matrix from the dye matrix to create a matrix that was equal to the dye matrix, except that it had removed any areas where CAVIN was found on the other slide. Since, as stated above, CAVIN presence tended to exclude the areas near the cell membrane, this produced a reasonable first approximation to find the membranes of the cells in the slide.
This matrix was then run through the cv2.equalize histogram function. The function increases the contrast in an image by first creating a histogram of all of the pixel values in an image. It then finds cutoff points in the image where there are few to no pixels of a certain value (0 to lower bound and upper bound to 255) and removes these bounds from the image. The remaining nonzero part of the histogram is then stretched to cover the entire range of possible pixel values (0 to 255). These representative changes are also done in the image (moving each pixel in a direction away from the value of the mean in a manner proportional to the variance of the numerical pixel values of an image). This serves to magnify the contrast of the cell membranes much more than the remains of the dye, which, as a smooth gradient, do not pop out nearly as much. Then, further measures can be used to increase signal and reduce noise, which can include a low brightness filter and/or edge detection transforms and overlays.
The newly produced image has highlighted cell membrane structures, though these are riddled with discontinuities and interfering elements. These provide a rough outline of the various cell membranes, which can be used to reconstruct a complete structure by connecting them with other discontinuities. An approach that yielded promising results, and the one I picked, was to use convolutive morphological processing techniques. These involve creating a base matrix, or kernel, of a certain size, with certain elements that make them up. Each element in the kernel is then multiplied by a corresponding element in a matrix sized region of the image. These are all summed up and assigned to a particular pixel in the post processed image. Of these techniques, I used the “opening” transformation (erosion followed by dilation) to remove interfering bright spots in the dye layer. This is then followed by a dilation with a kernel of variable dimension to connect the cell membranes in the various slides. The high flow slides have cells and membranes that follow the direction of flow more closely, and thus can be connected with a high degree of accuracy if a horizontal kernel is used. The low flow slides, however, do not have an overwhelming directional influence, so a more rectangular kernel with high entropy elements must be used to segment the cells. Together, these various procedures create a serviceable template that can be further processed in the segmentation pipeline.