Project #3: Streamlining floodfill for convex constructs
Various streamlining techniques could be used to drastically reduce the computational load of running the floodfill, at a slight expense for characterizing edge cases. However, since the nucleus is a topologically homogeneous structure throughout all of its occurrences in the slide, this does not effectively reduce the accuracy in gathering data. Since the first algorithm linearly traces the entirety of the inside of the nucleus, it loops through the height of the nucleus while looping through the length of the nucleus, which results in a worst case time of O(~pi*height*width) for an ellipse. The new variant that makes use of the fact that the nucleus is a smooth and continuous structure. That is, an edge between the inside and outside of the structure is more likely to be smooth without large jumps in continuity and sharp angles. Using this, I first used the existing vertcaps function to find the top and bottom of the nucleus for a certain x value. After this, I ran a loop to find the next edge value using the top and bottom edge values as reference instead of the center of the nucleus, while storing the new bottom and top values for each iteration. This measures the circumference instead of the area to reduce the time from O(~pi*height*width) to O(height + circumference) for an ellipse. Circumference can be approximated in the following function, where c ~= 2*pi*sqrt((height^2 + width^2)/2) where c is a pseudolinear function. The linear regression axial finder is then used to quantify the orientation.